backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-generativeai
openpyxl
#python -m uvicorn app.main:app --reload --host 0.0.0.0 --port 8000
